move rules
pawn and knight move follow the rules (mostly)

pawn can move one square forward
it can move two squares on its first turn
it's set to only allow if it's on the second rank for white and the seventh for black. I don't see a reason a pawn would ever be on their starting ranks if they haven't moved, but I suppose this method isn't technically failsafed.
It can move diagonally forward one square if there is an enemy piece  in that square

the knight can move either two squares vertically and one horizontally or one square vertically and two squares horizontally in any direction

pawns cannot land on squares with pieces of the same color. knights can for the moment

there is no actual capture function and when two pieces 'exist' on the same square, the first one there defaults