bishop, more like bitchup
the rules for bishop moves are a bit more complicated than i anticipated.

detailing the squares they are allowed to move to is relatively straight forward. a chessboard is a grid that uses algebraic notation. bishops move diagonally, but it's technically up 1 and left 1 or up down on and right 1, and so on. This means that bishop moves must have a slope of either 1 or -1

 the tricky bit is getting the slope but then also checking each square between the start and end points to see if a piece occupies it.

with the simple slope method, the number of squares moved and the direction of movement are irrelevant, so long as the slope is 1 or -1.

With the method checking if each square in between is occupied, I have to calculate which squares are between the start and finish while also checking for pieces. For some reason I tried nesting three loops that ran on every click of a square. naturally this crashed the browser.

Here's the idea I'm working on. Splitting both slope types base on whether the rise is positive or negative. that way it will only check in the direction of the move. then I have to loop the coordinates through the list of pieces to see if it's occupied. if they aren't, the bishop can move

i also need to change the deselection process so I can click on an opposite colored piece without crashing