the bishop moves
the bishop can move in any legal direction and it won't jump over any other piece

I am going to make the function dryer. I was just happy to get it working

invalid move clicks deselect active piece and squares. this means more clicks are spent selecting and not going through the move logic. i also took out the images of chess pieces and replaced them with unicode

I also deleted their files and whatnot so the aren't part of the bundle

I am also going to consolidate the piece states. I don't see much of a reason to maintain two, but that might prove wrong

i need to add capturing for pieces and en passant for pawns. I don't anticipate the rook taking nearly as long, and for the most part, the queen is just a rook combined with a bishop

the actual game logic is to come.  i'm trying to figure out how to calculate if the king is in check both before and after a move

I originally made piece components, thinking i was going to use an object for every piece, but I've since learned that it's all in the squares. I've deleted all reference to them(that I know of)

with all of these changes, the app doesn't seem to freeze anymore