Rooks and functions
I created a function to group the functions for unselecting the active piece, origin square, and destination together for readability and to pass them more easily

I created a single function to check if squares are occupied by pieces. That is part of my plan to make the bishop rules easier to deal with. the other part is creating a function to deal with adding and subtracting to make checking legal moves more general

I'm still working on rook moves. i've gotten two directions. I'm trying to add the piece square checker into it