Legal at last
Most moves legal moves are allowed to be made and most illegal moves aren't.

the two main prloblems currently are en passant and check. castling is an issue as well, but I can't be taken care of until my check mechanics are straight

Problem with check:
The program can tell if a move is legal and make said move, however, it can't tell the rest of the program whether the move is legal or not. the checkLegalMoves module should take care of that, but retrofitting everything is difficult.

I'm making a new branch that sets out to do just that

Also need to add comments. It's just been me dealing with it, but it's getting to an unwieldy point